+++ Changelog.md
@@ -5,6 +5,7 @@
 - Fix: Error involving SecurityCraft's creative tabs when running SecurityCraft alongside certain other mods
 - Fix: Duplication exploit involving Reinforced Pistons
 - Fix: The name of certain SecurityCraft blocks is displayed within brackets by WTHIT and similar mods
+- Fix: Using bone meal on a Reinforced Grass Block does not grow plants on adjacent ones
 
 --------------------------Changelog for v1.9.8 of SecurityCraft--------------------------